+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Dates are formatted as year-month-day.
 - Added scan/purge buttons next to disk usage percentages
 
 ### Changes
+- Core power now matches the maximum number of core speed upgrades you have
 - Moved `Reset settings` to the same line as `Restart game` in the settings menu and made it permanently red
 
 ## v1.0.0 - 2020-09-08